fix: resolve pre-existing mypy errors and broken external links
Master-tip CI bitrot remaining after numpy/black fixes:
- mypy: annotate _get_annotator_quality as np.ndarray; assert cmap;
  cast array/Series inputs; int/float coercions for typed returns
- links: replace retired /projects and /slack URLs; ignore flaky
  stargazers badge URL in proof-html (bot 404s)

+++ cleanlab/datalab/internal/issue_manager/multilabel/label.py
@@ -120,7 +120,7 @@ class MultilabelIssueManager(IssueManager):
             },
         )
         # Get a summarized dataframe of the label issues
-        self.summary = self.make_summary(score=scores.mean())
+        self.summary = self.make_summary(score=float(scores.mean()))
 
         # Collect info about the label issues
         self.info = self.collect_info(self.datalab.labels, predicted_labels)
